requirements.txt
openai
pydantic
rich
python-dotenv

# Testing (optional)
pytest
pytest-asyncio
